#222 WIP def action_copy_paste and def manage_snapping moved to parent_init.py at the moment copy_paste works for ws_nodes and ud_nodes
