+++ parent_init.py
@@ -6,11 +6,11 @@ or (at your option) any later version.
 """
 
 # -*- coding: utf-8 -*-
-from qgis.core import QgsMapLayerRegistry
+from qgis.core import QgsMapLayerRegistry,  QgsExpression, QgsFeatureRequest, QgsPoint
 from qgis.utils import iface
 from qgis.gui import QgsMessageBar
 from PyQt4.Qt import QTableView, QDate
-from PyQt4.QtCore import QSettings, Qt
+from PyQt4.QtCore import QSettings, Qt, QPoint
 from PyQt4.QtGui import QLabel, QComboBox, QDateEdit, QPushButton, QLineEdit, QIcon, QWidget, QDialog, QTextEdit, QAction
 from PyQt4.QtGui import QSortFilterProxyModel, QCompleter, QStringListModel
 from PyQt4.QtSql import QSqlTableModel
@@ -1369,3 +1369,88 @@ class ParentDialog(QDialog):
         completer.setCompletionMode(QCompleter.UnfilteredPopupCompletion)
         widget.setCompleter(completer)
 
+
+    def action_copy_paste(self):
+        self.set_snapping()
+        self.emit_point.canvasClicked.connect(self.manage_snapping)
+
+
+    def manage_snapping(self, point):
+        # Get node of snapping
+        map_point = self.canvas.getCoordinateTransform().transform(point)
+        x = map_point.x()
+        y = map_point.y()
+        eventPoint = QPoint(x, y)
+
+        # Snapping
+        (retval, result) = self.snapper.snapToBackgroundLayers(eventPoint)  # @UnusedVariable
+
+        # That's the snapped point
+        if result:
+            for snapped_point in result:
+                if snapped_point.layer.name() == self.iface.activeLayer().name():
+                    # Get only one feature
+                    point = QgsPoint(snapped_point.snappedVertex)  # @UnusedVariable
+                    snapped_feature = next(
+                        snapped_point.layer.getFeatures(QgsFeatureRequest().setFilterFid(snapped_point.snappedAtGeometry)))
+                    snapped_feature_attr = snapped_feature.attributes()
+                    # Leave selection
+                    snapped_point.layer.select([snapped_point.snappedAtGeometry])
+                    break
+
+        aux = "\"node_id\" = "
+        aux += "'" + str(self.id) + "'"
+        expr = QgsExpression(aux)
+        if expr.hasParserError():
+            message = "Expression Error: " + str(expr.parserErrorString())
+            self.controller.show_warning(message)
+            return
+
+        layer = self.iface.activeLayer()
+        fields = layer.dataProvider().fields()
+
+        layer.startEditing()
+        it = layer.getFeatures(QgsFeatureRequest(expr))
+        id_list = [i for i in it]
+        # Sector, Dma, Exploitation, State, State type, Workcat, Builtdate, Verified, CATALOG
+        # sector_id, dma, expl_id, state, state_type, xxxxx_workcat_id, xxxxx_builddate, verified, nodecat_id
+        self.controller.log_info(str(id_list))
+        if id_list != []:
+            # id_list[0]: pointer on current feature
+            id_current = id_list[0].attribute('node_id')
+
+            message = "Selected snapped node to copy values from: " + str(snapped_feature_attr[0]) + "\n"
+            message += "Do you want to copy its values to the current node?\n\n"
+            # Replace id because we don't have to copy it!
+            snapped_feature_attr[0] = id_current
+            snapped_feature_attr_aux = []
+            fields_aux = []
+            for i in range(1, len(fields)):
+                if fields[i].name() == 'sector_id' or fields[i].name() == 'dma_id' or fields[i].name() == 'expl_id' or \
+                            fields[i].name() == 'state' or fields[i].name() == 'state_type' or fields[i].name() == \
+                            self.iface.activeLayer().name().lower()+'_workcat_id' or fields[i].name() == \
+                            self.iface.activeLayer().name().lower()+'_builtdate' or fields[i].name() == 'verified' or \
+                            fields[i].name() == 'nodecat_id' or fields[i].name() == 'arccat_id' or \
+                            fields[i].name() == 'connecat_id' or fields[i].name() == 'gullycat_id':
+                    snapped_feature_attr_aux.append(snapped_feature_attr[i])
+                    fields_aux.append(fields[i].name())
+                if self.project_type == 'ud':
+                    if fields[i].name() == 'node_type' or fields[i].name() == 'arc_type' or fields[i].name() == 'connec_type' \
+                            or fields[i].name() == 'gully_type':
+                        snapped_feature_attr_aux.append(snapped_feature_attr[i])
+                        fields_aux.append(fields[i].name())
+
+            for i in range(0, len(fields_aux)):
+                message += str(fields_aux[i]) + ": " + str(snapped_feature_attr_aux[i]) + "\n"
+
+            # if i==0:
+            #     snapped_feature_attr_aux.append(snapped_feature_attr[i])
+            # Show message before executing
+            answer = self.controller.ask_question(message, "Update records", None)
+            self.controller.log_info(str(message))
+            # If ok execute and refresh form
+            if answer:
+                id_list[0].setAttributes(snapped_feature_attr)
+                layer.updateFeature(id_list[0])
+                layer.commitChanges()
+                self.dialog.refreshFeature()
